one to many example

+++ src/main/resources/db/compositekey/cartitem/cart-item-schema.sql
@@ -1,7 +1,8 @@
 create TABLE cartitem (
-    cartid INTEGER NOT NULL,
-    productid INTEGER NOT NULL,
+    cartid BIGINT NOT NULL,
+    productid BIGINT NOT NULL,
     name VARCHAR(255) NOT NULL,
-    count INTEGER NOT NULL,
+    count BIGINT NOT NULL,
     PRIMARY KEY (cartid, productid)
 );
+
